Deprecation warning and conftest added and call dict removed from PredictMixin.
